dataPipeline/airflow_docker/requirements.txt
apache-airflow==2.5.3
flickrapi==2.4.0
pandas==1.3.4
